refactor: extract skill survey reconciliation helpers
